# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "BookShelf API",
        "description": "API REST para gerenciar uma estante de livros por gêneros.",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Servidor local"
        }
    ],
    "paths": {
...
        "/api/genres": {
            "get": {
                "tags": [
                    "Gêneros"
                ],
                "summary": "Listar gêneros",
                "description": "Retorna todos os gêneros cadastrados.",
                "operationId": "0ff78353b1c27ef8905e58f3cf31370a",
                "responses": {
                    "200": {
                        "description": "Lista de gêneros",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Genre"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Gêneros"
                ],
                "summary": "Criar gênero",
                "description": "Cria um novo gênero.",
                "operationId": "8d21057e41193b21373f4e2cce0a8c73",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name"
                                ],
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ficção",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Romances e histórias fictícias",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gênero criado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        },
        "/api/genres/{id}": {
            "get": {
                "tags": [
                    "Gêneros"
                ],
                "summary": "Exibir gênero",
                "description": "Retorna um gênero específico pelo ID.",
                "operationId": "cec9d00fae56cff69cba705f6b436983",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do gênero",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Gênero encontrado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Gênero não encontrado"
                    }
                }
            },
            "put": {
                "tags": [
                    "Gêneros"
                ],
                "summary": "Atualizar gênero",
                "description": "Atualiza um gênero existente pelo ID.",
                "operationId": "9ec4944f7fd50bd7280fb1366c748709",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do gênero",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ficção",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição atualizada",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gênero atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Gênero não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Gêneros"
                ],
                "summary": "Excluir gênero",
                "description": "Exclui um gênero pelo ID.",
                "operationId": "1890f1018746d6e4385aa0c3244acc08",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do gênero",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Gênero excluído"
                    },
                    "404": {
                        "description": "Gênero não encontrado"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Gêneros"
                ],
                "summary": "Atualizar parcialmente gênero",
                "description": "Atualiza parcialmente um gênero existente pelo ID.",
                "operationId": "267965d763a2c67745c6afab91e4a9f6",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID do gênero",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Ficção",
                                        "maxLength": 255
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Descrição atualizada",
                                        "nullable": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Gênero atualizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Genre"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Gênero não encontrado"
                    },
                    "422": {
                        "description": "Erro de validação"
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
...
            "Genre": {
                "required": [
                    "id",
                    "name",
                    "created_at",
                    "updated_at"
                ],
                "properties": {
                    "id": {
                        "type": "integer",
                        "example": 1
                    },
                    "name": {
                        "type": "string",
                        "example": "Ficção"
                    },
                    "description": {
                        "type": "string",
                        "example": "Romances e histórias fictícias",
                        "nullable": true
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-07T14:09:44.000000Z"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "example": "2026-04-07T14:09:44.000000Z"
                    }
                },
                "type": "object"
            }
        }
    },
...
}